Move ID field from DataJobEntity to JobEntity
