+++ src/pkg/caendr/caendr/models/datastore/gene_browser_tracks.py
@@ -26,7 +26,6 @@ class GeneBrowserTracks(JobEntity):
       *super().get_props_set(),
 
       # Submission
-      'id',
       'username',
 
       # Query
