+++ src/pkg/caendr/caendr/models/datastore/job_entity.py
@@ -2,6 +2,7 @@ from caendr.services.logger import logger
 
 from caendr.models.datastore import Container, Entity
 from caendr.models.datastore.pipeline_operation import PipelineOperation
+from caendr.utils.data import unique_id
 
 
 
@@ -38,13 +39,18 @@ class JobEntity(Entity):
     return super(JobEntity, cls).__new__(cls)
 
 
-  def __init__(self, *args, **kwargs):
+  def __init__(self, name_or_obj = None, *args, **kwargs):
 
     # Create Container object to store relevant Container fields
     self.__container = Container()
 
+    # If nothing passed for name_or_obj, create a new ID to use for this object
+    if name_or_obj is None:
+      name_or_obj = unique_id()
+      self.set_properties_meta(id = name_or_obj)
+
     # Initialize from superclass
-    super().__init__(*args, **kwargs)
+    super().__init__(name_or_obj, *args, **kwargs)
 
 
 
@@ -65,6 +71,26 @@ class JobEntity(Entity):
       'operation_name',
       'status',
     }
+
+
+  @classmethod
+  def get_props_set_meta(cls):
+    return {
+      *super().get_props_set_meta(),
+      'id',
+    }
+
+
+
+  ## Unique ID ##
+
+  @property
+  def id(self):
+    '''
+      This Entity's unique ID.
+      This field cannot be set manually; it must be determined at initialization.
+    '''
+    return self._get_meta_prop('id')
